+++ src/Views/Columns/Traits/IsColumn.php
@@ -5,7 +5,7 @@ namespace Rappasoft\LaravelLivewireTables\Views\Columns\Traits;
 use Rappasoft\LaravelLivewireTables\Traits\Core\HasLocalisations;
 use Rappasoft\LaravelLivewireTables\Views\Columns\Traits\Configuration\ColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Columns\Traits\Helpers\{ColumnHelpers};
-use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasAttributes, HasLabelAttributes, HasTheme};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\{HasAttributes, HasLabelAttributes, HasTheme, HasToolTip};
 
 trait IsColumn
 {
@@ -29,6 +29,7 @@ trait IsColumn
         HasLabelAttributes,
         HasSecondaryHeader,
         HasTheme,
+        HasToolTip,
         HasVisibility;
 
     // What displays in the columns header